Refactor drawer, body, subheader, footer
