+++ components/src/core/Drawer/Drawer.tsx
@@ -1,12 +1,13 @@
 import React, { useEffect, useState, useCallback, ReactNode } from 'react';
 import { createStyles, makeStyles, Theme, useTheme } from '@material-ui/core/styles';
-import { ButtonBaseProps as MuiButtonBaseProps, Drawer, DrawerProps } from '@material-ui/core';
+import { Drawer, DrawerProps } from '@material-ui/core';
 import { DrawerBodyProps } from './DrawerBody';
 import clsx from 'clsx';
 import PropTypes from 'prop-types';
-import { InfoListItemProps as BaseInfoListItemProps } from '../InfoListItem';
 import { useDrawerLayout } from '../DrawerLayout/contexts/DrawerLayoutContextProvider';
 import { DrawerContext } from './DrawerContext';
+import { NavItemSharedStyleProps, SharedStyleProps } from './types';
+import { mergeStyleProp } from './utilities';
 
 export const RAIL_WIDTH = 72;
 export const RAIL_WIDTH_CONDENSED = 56;
@@ -58,58 +59,6 @@ type DrawerClasses = {
     sideBorder?: string;
 };
 
-// type shared by Drawer, DrawerBody, DrawerNavGroup, NestedNavItem
-// these types are inherited from the Drawer level to the NestedNavItem
-// parent props will be overriden by the child props if defined
-export type PXBlueDrawerInheritableProperties = {
-    // Background color for the 'active' item
-    activeItemBackgroundColor?: string;
-
-    // shape of the active item background
-    activeItemBackgroundShape?: 'round' | 'square';
-
-    // Font color for the 'active' item
-    activeItemFontColor?: string;
-
-    // Icon color for the 'active' item
-    activeItemIconColor?: string;
-
-    // Whether to have chevrons for all menu items
-    chevron?: boolean;
-
-    // Icon used to collapse drawer
-    // default is expandIcon rotated 180 degrees
-    collapseIcon?: JSX.Element;
-
-    // Whether to show a line between all items
-    divider?: boolean;
-
-    // Icon used to expand drawer
-    expandIcon?: JSX.Element;
-
-    // Whether to hide the paddings reserved for menu item icons
-    hidePadding?: boolean;
-
-    // InfoListItem overrides for NavItem
-    InfoListItemProps?: Partial<BaseInfoListItemProps>;
-
-    // InfoListItem overrides for DrawerRailItem
-    ButtonBaseProps?: Partial<MuiButtonBaseProps>;
-
-    // The color used for the item text
-    itemFontColor?: string;
-
-    // The color used for the icon
-    itemIconColor?: string;
-
-    // internal API
-    // will apply to all menu items when onClick
-    onItemSelect?: () => void;
-
-    // Whether to apply material ripple effect to items
-    ripple?: boolean;
-};
-
 const findChildByType = (children: ReactNode, type: string): JSX.Element[] =>
     React.Children.map(children, (child: any) => {
         if (child && child.type) {
@@ -120,50 +69,39 @@ const findChildByType = (children: ReactNode, type: string): JSX.Element[] =>
         }
     }) || [];
 
-// type shared by Drawer, DrawerBody, DrawerNavGroup
-// inheritable props but not for NestedNavItem
-export type PXBlueDrawerNavGroupInheritableProperties = {
-    // itemID for the 'active' item
-    activeItem?: string;
+export type DrawerComponentProps = Omit<DrawerProps, 'translate' | 'variant'> &
+    SharedStyleProps &
+    NavItemSharedStyleProps & {
+        // the id for the currently active item
+        activeItem?: string;
 
-    // If true, disable semi-bold title styling for the active item's parents in the drawer hierarchy
-    disableActiveItemParentStyles?: boolean;
+        // custom classes for default style overrides
+        classes?: DrawerClasses;
 
-    // background color for nested menu items
-    nestedBackgroundColor?: string;
+        // Sets a smaller width when the drawer is using the rail variant
+        condensed?: boolean;
 
-    // Whether to show a line between nested menu items
-    nestedDivider?: boolean;
+        // Describes if this Drawer is used outside of a DrawerLayout
+        noLayout?: boolean;
 
-    // Font color for group header
-    titleColor?: string;
-} & PXBlueDrawerInheritableProperties;
+        // Function called whenever a navigation item or rail item is clicked
+        onItemSelect?: (id: string) => void;
 
-export type DrawerComponentProps = {
-    classes?: DrawerClasses;
+        // Controls the open/closed state of the drawer
+        open: boolean;
 
-    // Sets a smaller width when the drawer is using the rail variant
-    condensed?: boolean;
+        // Enables Drawer to automatically open on hover for persistent variants.
+        openOnHover?: boolean;
 
-    // Describes if this Drawer is used outside of a DrawerLayout
-    noLayout?: boolean;
+        // Toggles the drawer side border instead of a drop shadow
+        sideBorder?: boolean;
 
-    // Controls the open/closed state of the drawer
-    open: boolean;
+        // Drawer variant type
+        variant?: 'persistent' | 'permanent' | 'temporary' | 'rail';
 
-    // Enables Drawer to automatically open on hover for persistent variants.
-    openOnHover?: boolean;
-
-    // Toggles the drawer side border instead of a drop shadow
-    sideBorder?: boolean;
-
-    // Drawer variant type
-    variant?: 'persistent' | 'permanent' | 'temporary' | 'rail';
-
-    // Sets the width of the drawer (in px) when open
-    width?: number;
-} & PXBlueDrawerNavGroupInheritableProperties &
-    Omit<DrawerProps, 'translate' | 'variant'>;
+        // Sets the width of the drawer (in px) when open
+        width?: number;
+    };
 
 const DrawerRenderer: React.ForwardRefRenderFunction<unknown, DrawerComponentProps> = (
     props: DrawerComponentProps,
@@ -175,37 +113,35 @@ const DrawerRenderer: React.ForwardRefRenderFunction<unknown, DrawerComponentPro
     const { setPadding, setDrawerOpen } = useDrawerLayout();
     const [hover, setHover] = useState(false);
     const {
-        activeItem,
+        // Inheritable Props
         activeItemBackgroundColor,
         activeItemBackgroundShape,
         activeItemFontColor,
         activeItemIconColor,
         chevron,
-        classes,
         collapseIcon,
-        /* eslint-disable @typescript-eslint/no-unused-vars */
-        condensed,
         disableActiveItemParentStyles,
-        /* eslint-enable @typescript-eslint/no-unused-vars */
         divider,
         expandIcon,
         hidePadding,
-        InfoListItemProps,
-        ButtonBaseProps,
         itemFontColor,
         itemIconColor,
         nestedBackgroundColor,
         nestedDivider,
+        ripple,
+        // Drawer-specific props
+        activeItem,
+        classes,
+        condensed,
         noLayout = false,
         open,
         openOnHover,
         onItemSelect,
-        ripple,
         sideBorder = false,
-        titleColor,
         variant: variantProp,
         width,
-        ...drawerProps // for Material-UI's Drawer component
+        // Other MUI Drawer Props
+        ...drawerProps
     } = props;
 
     const variant = variantProp || 'persistent'; // to allow drawerLayout to override this
@@ -229,8 +165,8 @@ const DrawerRenderer: React.ForwardRefRenderFunction<unknown, DrawerComponentPro
         (): JSX.Element[] =>
             findChildByType(props.children, 'DrawerSubheader')
                 .slice(0, 1)
-                .map((child) => React.cloneElement(child, { drawerOpen: isDrawerOpen() })),
-        [isDrawerOpen, props.children]
+                .map((child) => React.cloneElement(child)),
+        [props.children]
     );
 
     const getBody = useCallback(
@@ -239,55 +175,50 @@ const DrawerRenderer: React.ForwardRefRenderFunction<unknown, DrawerComponentPro
                 .slice(0, 1)
                 .map((child) =>
                     React.cloneElement(child, {
-                        activeItem,
-                        activeItemBackgroundColor,
-                        activeItemFontColor,
-                        activeItemIconColor,
-                        activeItemBackgroundShape,
-                        chevron,
-                        collapseIcon,
-                        divider,
-                        expandIcon,
-                        hidePadding,
-                        InfoListItemProps,
-                        ButtonBaseProps,
-                        itemFontColor,
-                        itemIconColor,
-                        nestedBackgroundColor,
-                        nestedDivider,
-                        ripple,
-                        titleColor,
-                        drawerOpen: isDrawerOpen(),
-                        onItemSelect: () => {
-                            if (onItemSelect) {
-                                onItemSelect();
-                            }
-                            setHover(false);
-                        },
+                        // Inherited Props
+                        activeItemBackgroundColor: mergeStyleProp(
+                            activeItemBackgroundColor,
+                            child.props.activeItemBackgroundColor
+                        ),
+                        activeItemBackgroundShape: mergeStyleProp(
+                            activeItemBackgroundShape,
+                            child.props.activeItemBackgroundShape
+                        ),
+                        activeItemFontColor: mergeStyleProp(activeItemFontColor, child.props.activeItemFontColor),
+                        activeItemIconColor: mergeStyleProp(activeItemIconColor, child.props.activeItemIconColor),
+                        chevron: mergeStyleProp(chevron, child.props.chevron),
+                        collapseIcon: mergeStyleProp(collapseIcon, child.props.collapseIcon),
+                        disableActiveItemParentStyles: mergeStyleProp(
+                            disableActiveItemParentStyles,
+                            child.props.disableActiveItemParentStyles
+                        ),
+                        divider: mergeStyleProp(divider, child.props.divider),
+                        expandIcon: mergeStyleProp(expandIcon, child.props.expandIcon),
+                        hidePadding: mergeStyleProp(hidePadding, child.props.hidePadding),
+                        itemFontColor: mergeStyleProp(itemFontColor, child.props.itemFontColor),
+                        itemIconColor: mergeStyleProp(itemIconColor, child.props.itemIconColor),
+                        nestedBackgroundColor: mergeStyleProp(nestedBackgroundColor, child.props.nestedBackgroundColor),
+                        nestedDivider: mergeStyleProp(nestedDivider, child.props.nestedDivider),
+                        ripple: mergeStyleProp(ripple, child.props.ripple),
                     } as DrawerBodyProps)
                 ),
         [
-            activeItem,
             activeItemBackgroundColor,
+            activeItemBackgroundShape,
             activeItemFontColor,
             activeItemIconColor,
-            activeItemBackgroundShape,
             chevron,
             collapseIcon,
+            disableActiveItemParentStyles,
             divider,
             expandIcon,
             hidePadding,
-            InfoListItemProps,
-            ButtonBaseProps,
             itemFontColor,
             itemIconColor,
             nestedBackgroundColor,
             nestedDivider,
             ripple,
-            titleColor,
-            isDrawerOpen,
             onItemSelect,
-            setHover,
             props.children,
         ]
     );
@@ -296,8 +227,8 @@ const DrawerRenderer: React.ForwardRefRenderFunction<unknown, DrawerComponentPro
         (): JSX.Element[] =>
             findChildByType(props.children, 'DrawerFooter')
                 .slice(0, 1)
-                .map((child) => React.cloneElement(child, { drawerOpen: isDrawerOpen() })),
-        [isDrawerOpen, props.children]
+                .map((child) => React.cloneElement(child)),
+        [props.children]
     );
 
     const getDrawerContents = useCallback(
@@ -420,7 +351,6 @@ export const PXBlueDrawerNavGroupInheritablePropertiesPropTypes = {
     nestedDivider: PropTypes.bool,
     onItemSelect: PropTypes.func,
     titleColor: PropTypes.string,
-    ...PXBlueDrawerInheritablePropertiesPropTypes,
 };
 
 // @ts-ignore
@@ -436,7 +366,6 @@ DrawerComponent.propTypes = {
     sideBorder: PropTypes.bool,
     variant: PropTypes.oneOf(['persistent', 'permanent', 'temporary', 'rail']),
     width: PropTypes.number,
-    ...PXBlueDrawerNavGroupInheritablePropertiesPropTypes,
 };
 DrawerComponent.defaultProps = {
     classes: {},
