+++ components/src/core/Drawer/DrawerSubheader.tsx
@@ -1,9 +1,9 @@
 import React, { HTMLAttributes } from 'react';
 import { Divider } from '@material-ui/core';
-import PropTypes from 'prop-types';
+// import PropTypes from 'prop-types';
+import { useDrawerContext } from './DrawerContext';
 
 export type DrawerSubheaderProps = HTMLAttributes<HTMLDivElement> & {
-    drawerOpen?: boolean;
     divider?: boolean;
     hideContentOnCollapse?: boolean;
 };
@@ -12,7 +12,8 @@ const DrawerSubheaderRender: React.ForwardRefRenderFunction<unknown, DrawerSubhe
     props: DrawerSubheaderProps,
     ref: any
 ) => {
-    const { children, drawerOpen, divider = true, hideContentOnCollapse = true, ...otherDivProps } = props;
+    const { children, divider = true, hideContentOnCollapse = true, ...otherDivProps } = props;
+    const { open: drawerOpen } = useDrawerContext();
     return (
         <>
             <div
@@ -30,6 +31,5 @@ export const DrawerSubheader = React.forwardRef(DrawerSubheaderRender);
 
 DrawerSubheader.displayName = 'DrawerSubheader';
 
-DrawerSubheader.propTypes = {
-    drawerOpen: PropTypes.bool,
-};
+// TODO FIX ME
+DrawerSubheader.propTypes = {};
